requirements.txt
flask>=3.0
python-dotenv==1.2.2
pytz==2026.1.post1
requests>=2.32
